+++ stats.h
@@ -11,7 +11,7 @@
 
 
 void stats_inc_counter(uint64_t *const p);
-void stats_add_counter(uint64_t *const p, const uint16_t value);
+void stats_add_counter(uint64_t *const p, const uint64_t value);
 void stats_set(uint64_t *const p, const uint64_t value);
 void stats_add_average(uint64_t *const p, const int value);
